requirements.txt
flask==2.1.3
flask-migrate==3.1.0
flask-redis==0.4.0
flask-script==2.0.6
flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.3
pyjwt==2.4.0
